+++ src/main/java/gtexpert/integration/binnies/extrabees/ExtraBeesModule.java
@@ -1,0 +1,29 @@
+package gtexpert.integration.binnies.extrabees;
+
+import net.minecraft.item.crafting.IRecipe;
+import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+
+import gtexpert.api.GTEValues;
+import gtexpert.api.modules.GTEModule;
+import gtexpert.api.util.Mods;
+import gtexpert.integration.GTEIntegrationSubmodule;
+import gtexpert.integration.binnies.extrabees.recipes.ExtraBeesItemsRecipe;
+import gtexpert.modules.GTEModules;
+
+@GTEModule(
+           moduleID = GTEModules.MODULE_EXBEES,
+           containerID = GTEValues.MODID,
+           modDependencies = { Mods.Names.FORESTRY, Mods.Names.EXTRA_BEES },
+           name = "GTExpert Extra Bees(Binnie's Mods) Integration",
+           description = "Extra Bees(Binnie's Mods) Integration Module")
+public class ExtraBeesModule extends GTEIntegrationSubmodule {
+
+    @Override
+    public void postInit(FMLPostInitializationEvent event) {
+        ExtraBeesItemsRecipe.init();
+    }
+
+    @Override
+    public void registerRecipesNormal(RegistryEvent.Register<IRecipe> event) {}
+}