+++ src/main/java/gtexpert/integration/binnies/genetics/GeneticsModule.java
@@ -1,0 +1,29 @@
+package gtexpert.integration.binnies.genetics;
+
+import net.minecraft.item.crafting.IRecipe;
+import net.minecraftforge.event.RegistryEvent;
+import net.minecraftforge.fml.common.event.FMLPostInitializationEvent;
+
+import gtexpert.api.GTEValues;
+import gtexpert.api.modules.GTEModule;
+import gtexpert.api.util.Mods;
+import gtexpert.integration.GTEIntegrationSubmodule;
+import gtexpert.integration.binnies.genetics.recipes.GeneticsItemsRecipe;
+import gtexpert.modules.GTEModules;
+
+@GTEModule(
+           moduleID = GTEModules.MODULE_GENETICS,
+           containerID = GTEValues.MODID,
+           modDependencies = { Mods.Names.FORESTRY, Mods.Names.GENETICS },
+           name = "GTExpert Genetics(Binnie's Mods) Integration",
+           description = "Genetics(Binnie's Mods) Integration Module")
+public class GeneticsModule extends GTEIntegrationSubmodule {
+
+    @Override
+    public void postInit(FMLPostInitializationEvent event) {
+        GeneticsItemsRecipe.init();
+    }
+
+    @Override
+    public void registerRecipesNormal(RegistryEvent.Register<IRecipe> event) {}
+}